# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Game Store API",
    "description": "A complete game store API with customers, products, and invoices management",
    "version": "1.0.0",
    "contact": {
      "name": "Game Store API Support",
      "email": "support@gamestore.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/products/category": {
      "get": {
        "summary": "Get products by category",
        "description": "Get products filtered by category",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Category is required"
          }
        }
      }
    },
    "/api/products/platform": {
      "get": {
        "summary": "Get products by platform",
        "description": "Get products filtered by platform",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product platform"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products for platform",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Platform is required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique product identifier"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "platform": {
            "type": "string",
            "description": "Product platform"
          },
          "stock": {
            "type": "integer",
            "description": "Available stock quantity"
          },
          "active": {
            "type": "boolean",
            "description": "Product active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product last update timestamp"
          }
        },
        "required": ["id", "name", "description", "price", "created_at", "updated_at"]
      },
...
    }
  },
...
}